use node-webkit's cancel-replace-dialog after selecting an existing file

+++ app/view/main/MainController.js
@@ -188,31 +188,36 @@ Ext.define('Get.view.main.MainController', {
 			filename += '.get';
 		}
 
-		if (fs.existsSync(filename)) {
+		if (fs.existsSync(filename) && filename !== currentFilename) {
 			// TODO: Übernimmt Chrome in Windows auch die Überprüfung auf existierende Datei?
-			Ext.Msg.show({
-				message: '"' + path.basename(filename) + '" existiert bereits. Soll die Datei ersetzt werden?',
-				buttons: Ext.Msg.OKCANCEL,
-				buttonText: {
-					ok: 'Ersetzen',
-					cancel: 'Abbrechen'
-				},
-				icon: Ext.Msg.WARNING,
-				fn: function(choice) {
-					if (choice == 'ok') {
-						if (filename !== currentFilename) {
-							shell.rm(filename);
-						}
-						duplicate();
-					}
-				}
-			});
-			Ext.Msg.down('button#ok').addCls('btn-ok');
-			Ext.Msg.down('toolbar').setLayout({pack: 'end'});
+			shell.rm(filename);
 		}
-		else {
-			duplicate();
-		}
+		duplicate();
+
+		// if (fs.existsSync(filename)) {
+		// 	Ext.Msg.show({
+		// 		message: '"' + path.basename(filename) + '" existiert bereits. Soll die Datei ersetzt werden?',
+		// 		buttons: Ext.Msg.OKCANCEL,
+		// 		buttonText: {
+		// 			ok: 'Ersetzen',
+		// 			cancel: 'Abbrechen'
+		// 		},
+		// 		icon: Ext.Msg.WARNING,
+		// 		fn: function(choice) {
+		// 			if (choice == 'ok') {
+		// 				if (filename !== currentFilename) {
+		// 					shell.rm(filename);
+		// 				}
+		// 				duplicate();
+		// 			}
+		// 		}
+		// 	});
+		// 	Ext.Msg.down('button#ok').addCls('btn-ok');
+		// 	Ext.Msg.down('toolbar').setLayout({pack: 'end'});
+		// }
+		// else {
+		// 	duplicate();
+		// }
 	},
 
 	onCloseMenuItem: function() {
